requirements.txt
pyannote.audio==3.3.2
python-dotenv
speechbrain==1.0.2
audio-separator==0.24.1
onnxruntime==1.20.1
yt-dlp
groq==0.13.0
numba==0.60.0
ffmpeg-python
gruut==2.2.3
encodec==0.1.1
gruut_lang_en==2.0.0
gruut_lang_de==2.0.1
gruut_lang_fr==2.0.2
gruut_lang_es==2.0.1
umap-learn==0.5.1
anyascii==0.3.2
coqpit==0.0.17
einops==0.8.0
g2pkk==0.1.2
inflect==7.4.0
num2words==0.5.13
pysbd==0.3.4
trainer==0.0.36
Unidecode==1.3.8
pypinyin==0.53
dateparser==1.1.8
jsonlines==1.2.0
networkx==2.8.8
pynndescent==0.5.13
python_crfsuite==0.9.11
SudachiDict_core
SudachiPy==0.6.9
jamo
g2pkk
bangla
bnnumerizer
bnunicodenormalizer
dateparser
gruut-ipa
python_crfsuite
pynndescent
tzlocal
hangul_romanize
ascii-magic
tensorflow==2.8.0rc0
jieba
grpcio-status==1.33.2 
scikit-image
nltk==3.2.4
gradio==5.8.0
google-cloud-texttospeech==2.16.3